% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_spec.R
\name{as_survey}
\alias{as_survey}
\alias{as_survey.inzsvyspec}
\title{as_survey method}
\usage{
\method{as_survey}{inzsvyspec}(.data, ...)
}
\arguments{
\item{.data}{an \code{inzsvyspec} object}

\item{...}{additional arguments, ignored}
}
\value{
a survey design object
}
\description{
Coerce an object to a survey design by extracting the survey object
}
