% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey.R, R/make_survey.R
\name{import_survey}
\alias{import_survey}
\alias{make_survey}
\title{Import survey information from a file}
\usage{
import_survey(file, data, read_fun, ...)

make_survey(.data, spec)
}
\arguments{
\item{file}{the file containing survey information (see Details)}

\item{data}{optional, if supplied the survey object will be created with the supplied data.
Can be either a data.frame-like object, or a path to a data set which
will be imported using \code{iNZightTools::smart_read}.}

\item{read_fun}{function required to load the data specified in \code{file}}

\item{...}{additional arguments to \code{read_fun}}

\item{.data}{a data.frame}

\item{spec}{a \code{inzsvyspec} object}
}
\value{
a \code{inzsvyspec} object containing the design parameters and, if data supplied,
the created survey object
}
\description{
The survey information should be in TOML format, with fields
corresponding to survey design components. For example,\preformatted{strata = strata_var
clusters = cluster_var
weights = wt_var
}
}
\details{
For replicate weight designs, vectors (if necessary) are declared with
square brackets, like so:\preformatted{repweights = ['w01', 'w02', 'w03', 'w04', ..., 'w20']
}

although this would be better expressed using a regular expression,\preformatted{repweights = '^w[0-2]'
}

which matches all variables starting with a \code{w} followed by digits between 0 and 2 (inclusive).

Additionally, the information can contain a \code{file} specification
indicating the path to the data, which will be imported (if it exists
in the same directory as \code{file}) using \code{read_fun}, if specified;
or alternatively \code{file} can be a URL to a data file that will be downloaded
and read, again using \code{read_fun}.

If the data is not specified to \code{import_survey}, then \code{make_survey} can be used to manually
construct an \code{inzsvyspec} object with the design attached. This might be useful if you have
multiple datasets with the same design, for example.

\code{import_survey} calls \code{make_survey} when data is provided.
}
\section{Functions}{
\itemize{
\item \code{import_survey}: Import survey information from a file

\item \code{make_survey}: Construct a survey object from a data set and an \code{inzsvyspec} object
}}

\author{
Tom Elliott
}
