% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_survey.R
\name{make_survey}
\alias{make_survey}
\title{Make a survey object}
\usage{
make_survey(.data, spec)
}
\arguments{
\item{.data}{a data.frame}

\item{spec}{a \code{inzsvyspec} object}
}
\value{
a \code{inzsvyspec} object with the survey design loaded
}
\description{
Construct a survey object from a data set and an \code{inzsvyspec} object.
}
\author{
Tom Elliott
}
