% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spec.R
\name{write_spec}
\alias{write_spec}
\alias{write_spec.default}
\alias{write_spec.inzsvyspec}
\alias{write_spec.survey.design}
\title{Write a survey specification file}
\usage{
write_spec(x, file, data_path)

\method{write_spec}{default}(x, file, data_path)

\method{write_spec}{inzsvyspec}(x, file, data_path)

\method{write_spec}{survey.design}(x, file, data_path)
}
\arguments{
\item{x}{a survey object}

\item{file}{the path of a file to write to}

\item{data_path}{the location of the data, relative. Only specify if distributing both, and relative paths will remain consistent.}
}
\value{
Called for side-effect of writing to a file.
}
\description{
Writes a survey specification file for the given survey design object.
}
