% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_spec.R
\name{as_survey_spec}
\alias{as_survey_spec}
\alias{as_survey_spec.survey.design}
\title{Parse survey to survey spec}
\usage{
as_survey_spec(x, ...)

\method{as_survey_spec}{survey.design}(x, des, ...)
}
\arguments{
\item{x}{an object which can be converted to a survey spec (e.g., survey.design)}

\item{...}{additional arguments, used for methods}

\item{des}{a survey design object, needed for calibrated/poststratified surveys}
}
\value{
an \code{inzsvydesign} object, see \link{import_survey}
}
\description{
Parse survey to survey spec
}
\section{Methods (by class)}{
\itemize{
\item \code{as_survey_spec(survey.design)}: Method for survey.design objects

}}
